+++ pyimouapi/__init__.py
@@ -1,4 +1,4 @@
-__version__ = "1.2.9"
+__version__ = "1.3.0"
 
 from .device import ImouChannel, ImouDevice, ImouDeviceManager
 from .exceptions import (